raw: Add methods to send and receive msgs with Unix credentials
I have several projects that might make use of this crate, and one
of them today uses `SO_PEERCRED` to validate the sender's identity.

Add basic support for this.

I chose to make a small new struct for credentials
instead of exposing nix in our public types (perhaps we want to use
rustix in the future, etc.)

Closes: https://github.com/mitsuhiko/tokio-unix-ipc/issues/4