typed_channel: Make `channel` asymmetric, add `symmetric_channel`
To better handle the general case of asymmetric protocols.

I originally tried to just change `channel()` but it caused a lot
of pain in type inference in the test suite.  Change
most of the tests to use a `symmetric_channel()` as a convenient shorthand,
with one new test for an asymmetric `channel()`.