Drop `Sync` bound for channels
We shouldn't claim to support having multiple threads performing
operations on our channels at once.  Drop the explicit
`unsafe impl Sync` for our channel types.  Also,
by changing the `dead` variable to a `Cell`, this is an explicit
way to ensure `Sync` is *not* implemented.

Also, drop the `unsafe impl Send` because we get that by default
today, and we shouldn't be trying to override the auto traits
without good reason.

Our tests were all relying on this because the *default* for Tokio
is a multi-threaded reactor.
But in practice, there's a strong argument being made in
the community in some places that the default should really
be for non-`Send` futures, because it's actually pretty uncommon
for many workloads to truly benefit from work-stealing.

As this is a relatively low-level library, anyone who wants
multithreading can wrap our values in a `Mutex` for example.

This came up in the context of
https://github.com/mitsuhiko/tokio-unix-ipc/issues/16
and while it's still possible for two spawned tasks in a single